Start reverting some changes
