--- src/test/scala-3/com.scalatsi/DerivingTests.scala
@@ -1,16 +1,0 @@
-package com.scalatsi
-
-import TypescriptType.*
-
-import org.scalatest.flatspec.AnyFlatSpec
-import org.scalatest.matchers.should.Matchers
-
-class DerivingTests extends AnyFlatSpec with Matchers {
-  /*case class TestDerivingTSIType(foo: String) derives TSIType
-
-  "a case class" should "be able to derive a TSIType" in {
-    summon[TSIType[TestDerivingTSIType]] shouldBe TSType.interface("ITestDerivingTSIType", "foo" -> TSString)
-  }
-   */
-
-}
